fix(fix/homeandapi): created a separated file for api
