+++ Arquibancada Watch App/ArquibancadaApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct Arquibancada_Watch_AppApp: App {
     var body: some Scene {
         WindowGroup {
-            GameView()
+            ContentView()
         }
     }
 }
