+++ Arquibancada Watch App/GameView.swift
@@ -9,97 +9,58 @@ import SwiftUI
 
 
 struct GameView: View {
-    @State var away_team_en : String = ""
-    @State var away_score : Int = 0
-    @State var away_flag : String = ""
-    @State var home_team_en : String = ""
-    @State var home_score : Int = 0
-    @State var home_flag : String = ""
     
-    @State var urlStr = "http://api.cup2022.ir/api/v1/match/"
-    @State var yourToken = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MzdjZDgwNzQ4NzA5MjMzZmQ5ZWU1MzIiLCJpYXQiOjE2NjkyMzgzNTYsImV4cCI6MTY2OTMyNDc1Nn0.TB65_PmnS8mNE4wRcNafu1VKimecVkMYmKsxeiFX9IY"
-    @State var matchnum : Int = 3
+    @ObservedObject var api : API = API()
     
     var body: some View {
-        HStack{
-            
-            VStack{
-                
-                AsyncImage(url: URL(string: away_flag)) { image in
-                    image
-                        .resizable()
-                        .scaledToFill()
-                } placeholder: {
-                    ProgressView()
-                }
-                .frame(width: 64, height: 44)
-                .clipShape(RoundedRectangle(cornerRadius: 5))
-                Text(away_team_en)
-                Text("\(away_score)")
-            }
-            VStack{
-                AsyncImage(url: URL(string: home_flag)) { image in
-                    image
-                        .resizable()
-                        .scaledToFill()
-                } placeholder: {
-                    ProgressView()
-                }
-                .frame(width: 64, height: 44)
-                .clipShape(RoundedRectangle(cornerRadius: 5))
-                Text(home_team_en)
-                Text("\(home_score)")
-            }
-        }.task {
-            await loadData()
-        }
-    }
-    
-    func loadData() async {
-        guard let url = URL(string: urlStr + "\(matchnum)") else {
-            print("Invalid URL")
-            return
-        }
         
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        request.setValue("application/json", forHTTPHeaderField:"Content-Type")
-        request.setValue("Bearer \(yourToken)", forHTTPHeaderField:"Authorization")
-        request.timeoutInterval = 60.0
-        
-        let dataTask = URLSession.shared.dataTask(with: request) { (data, response, error) in
-            if let error = error {
-                print("Request error: ", error)
-                return
-            }
-            
-            if let httpResponse = response as? HTTPURLResponse{
-                if httpResponse.statusCode == 401{
-                    print("Refresh token...")
-                    return
+        VStack{
+            HStack(spacing: 5){
+                VStack(alignment: .center, spacing: 10){
+                    Group{
+                        AsyncImage(url: URL(string: api.match?.data[0].away_flag ?? "")) { image in
+                            image
+                                .resizable()
+                                .scaledToFill()
+                        } placeholder: {
+                            ProgressView()
+                        }
+                        .frame(width: 64, height: 44)
+                        .scaledToFill()
+                        .clipShape(RoundedRectangle(cornerRadius: 5))
+                        Text(api.match?.data[0].away_team_en.prefix(3).uppercased() ?? "TEAM 1").accessibilityLabel(api.match?.data[0].away_team_en ?? "Team one")
+                        Text("\(api.match?.data[0].away_score ?? 0)")
+                    }
+                }
+                
+                Image(systemName: "xmark").resizable().frame(width: 20, height: 20)
+                
+                VStack(alignment: .center, spacing: 10){
+                    Group{
+                        AsyncImage(url: URL(string: api.match?.data[0].home_flag ?? "")) { image in
+                            image
+                                .resizable()
+                                .scaledToFill()
+                        } placeholder: {
+                            ProgressView()
+                        }
+                        .frame(width: 64, height: 44)
+                        .scaledToFill()
+                        .clipShape(RoundedRectangle(cornerRadius: 5))
+                        Text(api.match?.data[0].home_team_en.prefix(3).uppercased() ?? "TEAM 2").accessibilityLabel(api.match?.data[0].away_team_en ?? "Team one")
+                        Text("\(api.match?.data[0].home_score ?? 0)")
+                    }
                 }
             }
             
-            guard let data = data else { return }
-            DispatchQueue.main.async {
-                
-                
-                if let decodedResponse = try? JSONDecoder().decode(Match.self, from: data){
-                    print(decodedResponse)
-                    self.away_team_en = decodedResponse.data[0].away_team_en
-                    self.away_score = decodedResponse.data[0].away_score
-                    self.away_flag = decodedResponse.data[0].away_flag
-                    self.home_team_en = decodedResponse.data[0].home_team_en
-                    self.home_score = decodedResponse.data[0].home_score
-                    self.home_flag = decodedResponse.data[0].home_flag
-                }
-                
-                //                let word = String(data: data, encoding: .utf8)
-                
+            Text("90' + 2")
+                .font(.footnote)
+        }.onAppear(){
+            Task.init {
+                await api.loadData(matchnum: 4)
             }
             
         }
-        dataTask.resume()
     }
 }
